+++ src/bluesearch/entrypoint/database/parse.py
@@ -25,7 +25,8 @@ from bluesearch.database.article import (
     Article,
     ArticleParser,
     CORD19ArticleParser,
-    PubmedXMLParser,
+    PMCXMLParser,
+    PubMedXMLParser,
 )
 
 
@@ -37,7 +38,7 @@ def get_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "article_type",
         type=str,
-        choices=("cord19-json", "pmc-xml"),
+        choices=("cord19-json", "pmc-xml", "pubmed-xml"),
         help="""
         Article format. If parsing several articles, all articles
         must have same format.
@@ -92,7 +93,9 @@ def run(
                 with inp.open() as f_inp:
                     parser_inst = CORD19ArticleParser(json.load(f_inp))
             elif article_type == "pmc-xml":
-                parser_inst = PubmedXMLParser(inp)
+                parser_inst = PMCXMLParser(inp)
+            elif article_type == "pubmed-xml":
+                parser_inst = PubMedXMLParser(inp)
             else:
                 raise ValueError(f"Unsupported article type {article_type}")