+++ docs/source/whatsnew.rst
@@ -29,6 +29,7 @@ Legend
 
 Latest
 ======
+- |Add| an :code:`ArticleParser` for metadata and abstracts from :code:`PubMed`.
 - |Change| the behaviour of :code:`bbs_database add` when no article was loaded
   from the given path. Now, stop with a :code:`RuntimeWarning` and don't load
   the NLP model to get sentences (fail faster).
@@ -39,7 +40,7 @@ Latest
   :code:`bbs_database parse`, and :code:`bbs_database add` to initialize
   a literature database, parse, and integrate articles.
 - |Add| research of topic at journal and article levels in :code:`topic` module.
-- |Add| :code:`PubmedXMLParser` to parse PubMed articles in XML JATS format.
+- |Add| :code:`PMCXMLParser` to parse PubMed articles in XML JATS format.
 - |Fix| DVC pipeline named :code:`sentence_embedding` regarding missing
   :code:`deps` elements and mixed models origin.
 - |Fix| the incorrect maximum input length to the transformer model used as